[API]: Support global github credentials - cleanups
